+++ ofdb-core/src/lib.rs
@@ -1,4 +1,4 @@
-use ofdb_entities::{category::*, email::*, event::*, nonce::*, place::*, user::*};
+use ofdb_entities::{address::*, category::*, email::*, event::*, nonce::*, place::*, user::*};
 
 pub trait EmailGateway {
     fn compose_and_send(&self, recipients: &[Email], subject: &str, body: &str);
@@ -18,4 +18,8 @@ pub trait NotificationGateway {
     fn user_registered_ofdb(&self, user: &User);
     fn user_registered(&self, user: &User, url: &str);
     fn user_reset_password_requested(&self, email_nonce: &EmailNonce);
+}
+
+pub trait GeoCodingGateway {
+    fn resolve_address_lat_lng(&self, addr: &Address) -> Option<(f64, f64)>;
 }