+++ ofdb-gateways/src/notify.rs
@@ -1,36 +1,23 @@
-use crate::{mailgun::Mailgun, sendmail::SendmailGateway, user_communication};
+use crate::user_communication;
 use ofdb_core::{EmailGateway, NotificationGateway};
 use ofdb_entities::{category::*, email::*, event::*, nonce::*, place::*, user::*};
-use std::env;
 
-const FROM_ADDRESS: &str = "\"Karte von morgen\" <no-reply@kartevonmorgen.org>";
-const DOMAIN: &str = "kartevonmorgen.org";
-
-pub struct NotificationGW {
+pub struct Notify {
     email_gw: Box<dyn EmailGateway + Send + Sync + 'static>,
 }
 
-impl NotificationGW {
-    pub fn new() -> Self {
-        match env::var("MAILGUN_API_KEY") {
-            Ok(api_key) => Self {
-                email_gw: Box::new(Mailgun {
-                    from_email: Email::from(FROM_ADDRESS),
-                    domain: DOMAIN.to_string(),
-                    api_key,
-                }),
-            },
-            Err(_) => {
-                warn!("No Mailgun API key found: Use sendmail as fallback");
-                Self {
-                    email_gw: Box::new(SendmailGateway::new(Email::from(FROM_ADDRESS))),
-                }
-            }
+impl Notify {
+    pub fn new<G>(gw: G) -> Self
+    where
+        G: EmailGateway + Send + Sync + 'static,
+    {
+        Self {
+            email_gw: Box::new(gw),
         }
     }
 }
 
-impl NotificationGateway for NotificationGW {
+impl NotificationGateway for Notify {
     fn place_added(
         &self,
         email_addresses: &[String],