+++ src/ports/cli.rs
@@ -1,12 +1,15 @@
 use crate::{
     core::prelude::*,
-    infrastructure::db::{sqlite, tantivy},
+    infrastructure::{
+        db::{sqlite, tantivy},
+        GEO_CODING_GW,
+    },
     ports::web,
 };
 
 use clap::{crate_authors, App, Arg};
 use dotenv::dotenv;
-use ofdb_gateways::opencage;
+use ofdb_core::GeoCodingGateway;
 use std::{env, path::Path};
 
 const DEFAULT_DB_URL: &str = "openfair.db";
@@ -19,7 +22,7 @@ fn update_event_locations<D: Db>(db: &mut D) -> Result<()> {
     for mut e in events {
         if let Some(ref mut loc) = e.location {
             if let Some(ref addr) = loc.address {
-                if let Some((lat, lng)) = opencage::resolve_address_lat_lng(addr) {
+                if let Some((lat, lng)) = GEO_CODING_GW.resolve_address_lat_lng(addr) {
                     if let Some(pos) = MapPoint::try_from_lat_lng_deg(lat, lng) {
                         if pos.is_valid() {
                             if let Err(err) = db.update_event(&e) {